Updated branding.
