+++ root/layout.tt
@@ -4,22 +4,14 @@
     <body>
         [%- IF c.debug -%][%- INCLUDE 'debug/stats.tt' -%][%- END ~%]
 
-        <div id="header">
-            <div id="header-logo">
-                <a href="/" class="logo" title="MusicBrainz"><strong>MusicBrainz</strong></a>
-                <div>
-                    [% USE sidebar = FormRenderer(sidebar_search, id_prefix => 'header') ~%]
-                    <form action="[% c.uri_for('/search') %]" method="get">
-                        [% sidebar.text('query', { placeholder => l('search') }) %]
-                        [% sidebar.select('type') %]
-                        [% sidebar.hidden('method', { value => 'indexed' }) %]
-                        [% form_submit(l('Search'), 'inline') %]
-                    </form>
+        <div class="header">
+            <a href="/" title="MusicBrainz"><img src="/static/images/layout/header-logo.svg" class="logo" /></a>
+            <div class="right">
+                <div class="top">
+                    [%- INCLUDE "layout/header/top.tt" -%]
                 </div>
-            </div>
-            <div id="header-menu">
-                <div>
-                    [%- INCLUDE "layout/menu.tt" -%]
+                <div class="bottom">
+                    [%- INCLUDE "layout/header/bottom.tt" -%]
                 </div>
             </div>
         </div>
