+++ root/static/styles/extra/banner.less
@@ -1,9 +1,11 @@
 .banner {
     margin: @form-margin 0;
+    margin-left: 12px;
+    margin-right: 12px;
     text-align: center;
-    background: #FFFFFF;
+    border: @medium-border 1px solid;
+    background: @effectively-white-bg;
     padding: @form-margin;
-    border: 0;
     position: relative;
 
     @media @wide {
